refactor(dashboard): modularize dashboard components for better maintainability
Move table and card components into separate files and extract layout logic into a reusable DashboardLayout component. This improves code organization and reusability across the dashboard.
